+++ cobblegen/src/main/java/io/github/null2264/cobblegen/CobbleGen.java
@@ -3,6 +3,7 @@ package io.github.null2264.cobblegen;
 import com.mojang.brigadier.CommandDispatcher;
 import com.mojang.brigadier.builder.LiteralArgumentBuilder;
 import io.github.null2264.cobblegen.compat.LoaderCompat;
+import io.github.null2264.cobblegen.compat.ModContainerCompat;
 import io.github.null2264.cobblegen.compat.TextCompat;
 import io.github.null2264.cobblegen.data.config.ConfigData;
 import io.github.null2264.cobblegen.data.config.ConfigMetaData;
@@ -13,6 +14,8 @@ import org.jetbrains.annotations.ApiStatus;
 
 import java.io.File;
 import java.nio.file.Path;
+import java.util.ArrayList;
+import java.util.regex.PatternSyntaxException;
 
 import static io.github.null2264.cobblegen.data.config.ConfigHelper.loadConfig;
 import static io.github.null2264.cobblegen.util.Constants.OP_LEVEL_GAMEMASTERS;
@@ -75,8 +78,69 @@ public class CobbleGen
         );
     }
 
-    public enum Channel {
-        PING,
-        SYNC,
+    @ApiStatus.Internal
+    public enum CreateSupport {
+        NONE,
+        O_FIVE_ONE_E,
+        O_FIVE_ONE_F,
+        SIX_O,  // FIXME: Properly support 6.0.x
+    }
+
+    /**
+     * Check Create mod version. Returns integer, in case Create introduced yet another breaking change
+     */
+    @ApiStatus.Internal
+    public static CobbleGen.CreateSupport getCreateSupport() {
+        if (CobbleGen.META_CONFIG.create.loadIntegration) return CobbleGen.CreateSupport.NONE;
+
+        #if MC<=11605
+        return CobbleGen.CreateSupport.NONE;  // We don't support create integration for MC1.16.5
+        #else
+        try {
+            String version = ModContainerCompat.fromLoader("create").getVersionString();
+            ArrayList<Integer> split = new ArrayList<>();
+            String patch = "a";
+            // version = 0.5.1-a-build.69
+            for (String s : version.split("\\.")) {
+                try {
+                    // 0.5.
+                    split.add(Integer.valueOf(s));
+                } catch (NumberFormatException exc) {
+                    // 1a or 1-a-build -> 1
+                    split.add(Integer.valueOf(s.substring(0, 1)));
+                    // -a-build -> '', a, build | a-build -> a, build
+                    String[] patchSplit = s.substring(1).split("-");
+                    if (patchSplit[0].equalsIgnoreCase("")) {
+                        patch = patchSplit[1];
+                    } else {
+                        // For older version that use 0.5.0a format instead of 0.5.0-a
+                        patch = patchSplit[0];
+                    }
+                    break;
+                }
+            }
+            if (split.get(0) == 0) {  // 0.x.x
+                #if FORGE
+                return CobbleGen.CreateSupport.O_FIVE_ONE_E;  // Always use Patch E mixin for Forge-alike
+                #else
+                if (split.get(1) == 5 && split.get(2) == 0 || split.get(1) < 5) return CobbleGen.CreateSupport.O_FIVE_ONE_E;
+                if (split.get(1) > 5 || split.get(2) > 1) return CobbleGen.CreateSupport.O_FIVE_ONE_F;  // Assume they don't introduce breaking changes on version bump
+
+                int compare = patch.compareToIgnoreCase("e");
+                if (compare <= 0) return CobbleGen.CreateSupport.O_FIVE_ONE_E;
+                return CobbleGen.CreateSupport.O_FIVE_ONE_F;
+                #endif
+            } else if (split.get(0) == 6) {  // 6.x.x
+                // TODO
+                CGLog.warn("Create mod v6+ is not yet supported");
+                return CobbleGen.CreateSupport.NONE;
+            }
+            // They're definitely going to introduce breaking changes on major version bump
+            return CobbleGen.CreateSupport.NONE;
+        } catch (java.util.NoSuchElementException | PatternSyntaxException | NumberFormatException exc) {
+            CGLog.error(exc);
+            return CobbleGen.CreateSupport.NONE;
+        }
+        #endif
     }
 }