+++ cobblegen/src/main/java/io/github/null2264/cobblegen/mixin/core/CobbleGenMixinPlugin.java
@@ -1,5 +1,6 @@
 package io.github.null2264.cobblegen.mixin.core;
 
+import io.github.null2264.cobblegen.CobbleGen;
 import io.github.null2264.cobblegen.compat.LoaderCompat;
 import io.github.null2264.cobblegen.util.CGLog;
 import org.objectweb.asm.tree.ClassNode;
@@ -26,68 +27,17 @@ public class CobbleGenMixinPlugin implements IMixinConfigPlugin
         return null;
     }
 
-    /**
-     * Check Create mod version. Returns integer, in case Create introduced yet another breaking change
-     *
-     * @return -1 for Unsupported, 0 for Patch E or older, 1 for Patch F or newer
-     */
-    private int getPatchVersion() {
-        #if FORGE
-        return 0;  // Always use Patch E mixin for Forge-alike
-        #elif MC<=11605
-        return -1;  // We don't support create integration for MC1.16.5
-        #else
-        try {
-            String version =
-                net.fabricmc.loader.api.FabricLoader.getInstance().getModContainer("create")
-                    .orElseThrow().getMetadata().getVersion().getFriendlyString();
-            ArrayList<Integer> split = new ArrayList<>();
-            String patch = "a";
-            // version = 0.5.1-a-build.69
-            for (String s : version.split("\\.")) {
-                try {
-                    // 0.5.
-                    split.add(Integer.valueOf(s));
-                } catch (NumberFormatException exc) {
-                    // 1a or 1-a-build -> 1
-                    split.add(Integer.valueOf(s.substring(0, 1)));
-                    // -a-build -> '', a, build | a-build -> a, build
-                    String[] patchSplit = s.substring(1).split("-");
-                    if (patchSplit[0].equalsIgnoreCase("")) {
-                        patch = patchSplit[1];
-                    } else {
-                        // For older version that use 0.5.0a format instead of 0.5.0-a
-                        patch = patchSplit[0];
-                    }
-                    break;
-                }
-            }
-            if (split.get(0) == 0) {
-                if (split.get(1) == 5 && split.get(2) == 0 || split.get(1) < 5) return 0;
-                if (split.get(1) > 5 || split.get(2) > 1) return 1;  // Assume they don't introduce breaking changes on version bump
-
-                int compare = patch.compareToIgnoreCase("e");
-                if (compare <= 0) return 0;
-                return 1;
-            }
-            // They're definitely going to introduce breaking changes on major version bump
-            return -1;
-        } catch (java.util.NoSuchElementException | PatternSyntaxException | NumberFormatException exc) {
-            CGLog.error(exc);
-            return -1;
-        }
-        #endif
-    }
-
     @Override
     public boolean shouldApplyMixin(String targetClassName, String mixinClassName) {
         if (mixinClassName.contains("CreateFluidReactionsMixin")) {
             if (!LoaderCompat.isModLoaded("create")) return false;
 
-            int patchVersion = getPatchVersion();
-            if (patchVersion == -1) return false;
-            if (mixinClassName.endsWith("PatchF")) return patchVersion >= 1;
-            if (mixinClassName.endsWith("PatchE")) return patchVersion <= 0;
+            CobbleGen.CreateSupport createSupport = CobbleGen.getCreateSupport();
+            if (createSupport == CobbleGen.CreateSupport.NONE) return false;
+            if (mixinClassName.endsWith("$OFive"))
+                return createSupport == CobbleGen.CreateSupport.O_FIVE_ONE_F || createSupport == CobbleGen.CreateSupport.O_FIVE_ONE_E;
+            if (mixinClassName.endsWith("$PatchF")) return createSupport == CobbleGen.CreateSupport.O_FIVE_ONE_F;
+            if (mixinClassName.endsWith("$PatchE")) return createSupport == CobbleGen.CreateSupport.O_FIVE_ONE_E;
         }
 
         #if MC>=12105