+++ cobblegen/src/main/java/io/github/null2264/cobblegen/mixin/create/CreateFluidReactionsMixin$PatchE.java
@@ -26,7 +26,7 @@ import static io.github.null2264.cobblegen.CobbleGen.FLUID_INTERACTION;
  */
 @Pseudo
 @Mixin(targets = {"com.simibubi.create.content.contraptions.fluids.FluidReactions", "com.simibubi.create.content.fluids.FluidReactions"})
-public abstract class CreateFluidReactionsMixinPatchE
+public abstract class CreateFluidReactionsMixin$PatchE
 {
     @Inject(
         method = "handlePipeFlowCollision",