+++ cobblegen/src/main/java/io/github/null2264/cobblegen/mixin/create/CreateFluidReactionsMixin$PatchF.java
@@ -18,7 +18,7 @@ import static io.github.null2264.cobblegen.CobbleGen.FLUID_INTERACTION;
  */
 @Pseudo
 @Mixin(targets = {"com.simibubi.create.content.contraptions.fluids.FluidReactions", "com.simibubi.create.content.fluids.FluidReactions"})
-public abstract class CreateFluidReactionsMixinPatchF
+public abstract class CreateFluidReactionsMixin$PatchF
 {
     @Inject(
         method = "handlePipeFlowCollisionFallback",