+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Fix crashes when being loaded alongside unsupported Create mod
+- Fix `cobblegen-meta.create.loadIntegration = false` didn't stop CobbleGen from trying to mixin Create mod
+
 ## [v5.4.0] - 2025-04-07
 
 ### Added